Rename snow_cli_global_context -> cli_global_context
