+++ src/snowcli/app/main_typer.py
@@ -2,7 +2,7 @@ import click
 from typer import Typer
 
 from snowcli.cli.common.flags import DEFAULT_CONTEXT_SETTINGS
-from snowcli.cli.common.snow_cli_global_context import global_context
+from snowcli.cli.common.cli_global_context import global_context
 
 
 def _handle_exception(exception: Exception):
