+++ src/snowcli/cli/common/experimental_behaviour.py
@@ -1,4 +1,4 @@
-from snowcli.cli.common.snow_cli_global_context import global_context
+from snowcli.cli.common.cli_global_context import global_context
 
 
 def experimental_behaviour_enabled() -> bool:
